Dearest Readers,
We regret to inform you that the enterprise
odd verifier team has forgotten an edge case,
and we had not created a test case for it before
the implementation was already complete. We see this
as nothing but an absolute failure to uphold our
own personal ideals.

When we first founded this company, we agreed that
we would enfore nothing but the most optimal Test
Driven Development practices. This commit message
you are reading today is proof that we have failed
to uphold this doctrine.

As a repentance for our sins, we have all agreed to
cease development of this repo entirely, and we will
recreate it from scratch with nothing but the most
rigorous testing and code review pipeline.

Thank you for being attentive readers of the
EnterpriseOddVerifier embargo.

Additionally, this commit message is atomic.